Use override_config in genesis
