+++ zk_toolbox/crates/config/Cargo.toml
@@ -18,6 +18,7 @@ ethers.workspace = true
 rand.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+serde_yaml.workspace = true
 strum.workspace = true
 thiserror.workspace = true
 types.workspace = true
